+++ Changelog.md
@@ -1,5 +1,10 @@
 ## Release History
 
+### 1.0.1 (2023-09-12)
+- [x] Segment operation `splitSegments` was added;
+- [x] Segment operation `joinSegments` was added;
+- [x] Segment operation `updateState` was added;
+
 ### 1.0.0 (2023-08-13)
  - [x] Works with 'dt-toolbox' version 7.x.x;